requirements.txt
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
azure-identity==1.16.0
azure-ai-ml
aiohttp
ipykernel
python-dotenv
requests
openai
SQLAlchemy
langchain
langchain-community
langchain-openai
python-dotenv
pyodbc
pandas